Reuse commandbutton logic
